Moved UserCredentialsDto to security package
